requirements.txt
# Used in producer and consumer Dockerfiles
fastavro
confluent-kafka
psycopg2-binary
flask
requests
avro-python3